requirements.txt
Pillow==9.2.0
rectpack==0.2.2
toml==0.10.2